srcs/requirements/backend/requirements.txt
asgiref==3.8.1
Django==5.1.5
django-formtools==2.5.1
django-otp==1.5.4
django-phonenumber-field==8.0.0
django-two-factor-auth==1.17.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
phonenumbers==8.13.54
psycopg==3.2.4
psycopg-binary==3.2.4
PyJWT==2.10.1
pypng==0.20220715.0
qrcode==7.4.2
sqlparse==0.5.3
typing_extensions==4.12.2
